# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/link/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Link"
                ],
                "summary": "Создать короткую ссылку по оригинальной",
                "parameters": [
                    {
                        "description": "OriginalLink",
                        "name": "link",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OriginalLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ссылка создана",
                        "schema": {
                            "$ref": "#/definitions/models.ShortCodeResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/link/{code}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Link"
                ],
                "summary": "Возвращает оригинальную ссылку по короткой, используя адрес короткой ссылки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Содержание короткой ссылки",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Оригинальная ссылка",
                        "schema": {
                            "$ref": "#/definitions/models.OriginalLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Ссылка не найдена",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.OriginalLinkRequest": {
            "type": "object",
            "properties": {
                "original_link": {
                    "type": "string"
                }
            }
        },
        "models.OriginalLinkResponse": {
            "type": "object",
            "properties": {
                "original_link": {
                    "type": "string"
                }
            }
        },
        "models.ShortCodeResponse": {
            "type": "object",
            "properties": {
                "short_code": {
                    "type": "string"
                },
                "short_link": {
                    "type": "string"
                }
            }
        }
    }
}